Rename Client to Conn
The functionality in Client has been merged into Conn.
Renamed a few fields in Conn so they're no longer exported and don't
collide with methods of the same name.
The contents of client_test.go were moved to conn_test.go.
Other misc types in client.go were moved to other source files.
