+++ CHANGELOG.md
@@ -10,20 +10,21 @@
 
 ### Breaking Changes
 * The minimum version of Go required to build this module is now 1.18.
+* The type `Client` has been renamed to `Conn`, and its constructor `New()` renamed to `NewConn()`.
 * Removed `ErrConnClosed` and `ErrTimeout` sentinel error types.
 * The following methods now require a `context.Context` as their first parameter.
-  * `Client.NewSession()`, `Session.NewReceiver()`, `Session.NewSender()`
+  * `Conn.NewSession()`, `Session.NewReceiver()`, `Session.NewSender()`
 * Removed `context.Context` parameter and `error` return from method `Receiver.Prefetched()`.
 * The following type names had the prefix `AMQP` removed to prevent stuttering.
   * `AMQPAddress`, `AMQPMessageID`, `AMQPSymbol`, `AMQPSequenceNumber`, `AMQPBinary`
 * Various `Default*` constants are no longer exported.
 * The args to `Receiver.ModifyMessage()` have changed.
-* The "variadic config" pattern for `Client`, `Session`, `Sender`, and `Receiver` constructors has been replaced with a struct-based config.
+* The "variadic config" pattern for `Conn`, `Session`, `Sender`, and `Receiver` constructors has been replaced with a struct-based config.
   * This removes the `ConnOption`, `SessionOption`, and `LinkOption` types and all of the associated configuration funcs.
   * The sender and receiver specific link options have been moved into their respective options types.
   * The `ConnTLS()` option was removed as part of this change.
 * The `Dial()` and `New()` constructors now require an `*ConnOptions` parameter.
-* `Client.NewSession()` now requires a `*SessionOptions` parameter.
+* `Conn.NewSession()` now requires a `*SessionOptions` parameter.
 * `Session.NewSender()` now requires `target` address and `*SenderOptions` parameters.
 * `Session.NewReceiver()` now requires `source` address and `*ReceiverOptions` parameters.
 * The various SASL configuration funcs have been slightly renamed.
@@ -36,10 +37,10 @@
 ### Bugs Fixed
 * Fixed potential panic in `muxHandleFrame()` when checking for manual creditor.
 * Fixed potential panic in `attachLink()` when copying source filters.
-* `New()` will no longer return a broken `*Client` in some instances.
+* `NewConn()` will no longer return a broken `*Conn` in some instances.
 * Incoming transfer frames received during initial link detach are no longer discarded.
 * Session will no longer flood peer with flow frames when half its incoming window is consumed.
-* Newly created `Session` won't leak if the context passed to `Client.NewSession()` expires before exit.
+* Newly created `Session` won't leak if the context passed to `Conn.NewSession()` expires before exit.
 * Newly created `link` won't leak if the context passed to `link.attach()` expires before exit.
 * Fixed an issue causing dispositions to hang indefinitely with batching enabled when the receiver link is disconnected.
