+++ link_test.go
@@ -406,7 +406,7 @@ func TestSessionFlowDisablesTransfer(t *testing.T) {
 	nextIncomingID := uint32(0)
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(SenderSettleModeUnsettled))
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
@@ -434,7 +434,7 @@ func TestSessionFlowDisablesTransfer(t *testing.T) {
 func TestExactlyOnceDoesntWork(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(SenderSettleModeUnsettled))
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
