+++ receiver_test.go
@@ -16,7 +16,7 @@ import (
 
 func TestReceiverInvalidOptions(t *testing.T) {
 	conn := mocks.NewNetConn(receiverFrameHandlerNoUnhandled(ReceiverSettleModeFirst))
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -69,7 +69,7 @@ func TestReceiverMethodsNoReceive(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -95,7 +95,7 @@ func TestReceiverMethodsNoReceive(t *testing.T) {
 
 func TestReceiverLinkSourceFilter(t *testing.T) {
 	conn := mocks.NewNetConn(receiverFrameHandlerNoUnhandled(ReceiverSettleModeFirst))
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -122,7 +122,7 @@ func TestReceiverLinkSourceFilter(t *testing.T) {
 
 func TestReceiverOnClosed(t *testing.T) {
 	conn := mocks.NewNetConn(receiverFrameHandlerNoUnhandled(ReceiverSettleModeFirst))
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -153,7 +153,7 @@ func TestReceiverOnClosed(t *testing.T) {
 
 func TestReceiverOnSessionClosed(t *testing.T) {
 	conn := mocks.NewNetConn(receiverFrameHandlerNoUnhandled(ReceiverSettleModeFirst))
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -184,7 +184,7 @@ func TestReceiverOnSessionClosed(t *testing.T) {
 
 func TestReceiverOnConnClosed(t *testing.T) {
 	conn := mocks.NewNetConn(receiverFrameHandlerNoUnhandled(ReceiverSettleModeFirst))
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -215,7 +215,7 @@ func TestReceiverOnConnClosed(t *testing.T) {
 
 func TestReceiverOnDetached(t *testing.T) {
 	conn := mocks.NewNetConn(receiverFrameHandlerNoUnhandled(ReceiverSettleModeFirst))
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -272,7 +272,7 @@ func TestReceiveInvalidMessage(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -394,7 +394,7 @@ func TestReceiveSuccessReceiverSettleModeFirst(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -453,7 +453,7 @@ func TestReceiveSuccessReceiverSettleModeSecondAccept(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -533,7 +533,7 @@ func TestReceiveSuccessReceiverSettleModeSecondAcceptOnClosedLink(t *testing.T) 
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -593,7 +593,7 @@ func TestReceiveSuccessReceiverSettleModeSecondReject(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -667,7 +667,7 @@ func TestReceiveSuccessReceiverSettleModeSecondRelease(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -746,7 +746,7 @@ func TestReceiveSuccessReceiverSettleModeSecondModify(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -846,7 +846,7 @@ func TestReceiveMultiFrameMessageSuccess(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -925,7 +925,7 @@ func TestReceiveInvalidMultiFrameMessage(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -1041,7 +1041,7 @@ func TestReceiveMultiFrameMessageAborted(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -1102,7 +1102,7 @@ func TestReceiveMessageTooBig(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -1148,7 +1148,7 @@ func TestReceiveSuccessAcceptFails(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -1213,7 +1213,7 @@ func TestReceiverDispositionBatcherTimer(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -1279,7 +1279,7 @@ func TestReceiverDispositionBatcherFull(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -1353,7 +1353,7 @@ func TestReceiverDispositionBatcherRelease(t *testing.T) {
 		}
 	}
 	conn := mocks.NewNetConn(responder)
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -1404,7 +1404,7 @@ func TestReceiverDispositionBatcherRelease(t *testing.T) {
 
 func TestReceiverCloseOnUnsettledWithPending(t *testing.T) {
 	conn := mocks.NewNetConn(receiverFrameHandlerNoUnhandled(ReceiverSettleModeFirst))
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -1431,7 +1431,7 @@ func TestReceiverCloseOnUnsettledWithPending(t *testing.T) {
 
 func TestReceiverConnReaderError(t *testing.T) {
 	conn := mocks.NewNetConn(receiverFrameHandlerNoUnhandled(ReceiverSettleModeFirst))
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
@@ -1465,7 +1465,7 @@ func TestReceiverConnReaderError(t *testing.T) {
 
 func TestReceiverConnWriterError(t *testing.T) {
 	conn := mocks.NewNetConn(receiverFrameHandlerNoUnhandled(ReceiverSettleModeFirst))
-	client, err := New(conn, nil)
+	client, err := NewConn(conn, nil)
 	require.NoError(t, err)
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	session, err := client.NewSession(ctx, nil)
