+++ session_test.go
@@ -41,7 +41,7 @@ func TestSessionClose(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 	for i := 0; i < 4; i++ {
 		ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -74,7 +74,7 @@ func TestSessionServerClose(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -114,7 +114,7 @@ func TestSessionCloseTimeout(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -131,7 +131,7 @@ func TestSessionCloseTimeout(t *testing.T) {
 func TestConnCloseSessionClose(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(SenderSettleModeUnsettled))
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -152,7 +152,7 @@ func TestConnCloseSessionClose(t *testing.T) {
 func TestSessionNewReceiverBadOptionFails(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(SenderSettleModeUnsettled))
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -196,7 +196,7 @@ func TestSessionNewReceiverBatchingOneCredit(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -239,7 +239,7 @@ func TestSessionNewReceiverBatchingEnabled(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -281,7 +281,7 @@ func TestSessionNewReceiverMismatchedLinkName(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -306,7 +306,7 @@ func TestSessionNewReceiverMismatchedLinkName(t *testing.T) {
 func TestSessionNewSenderBadOptionFails(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(SenderSettleModeUnsettled))
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -348,7 +348,7 @@ func TestSessionNewSenderMismatchedLinkName(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -370,7 +370,7 @@ func TestSessionNewSenderMismatchedLinkName(t *testing.T) {
 func TestSessionNewSenderDuplicateLinks(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(SenderSettleModeUnsettled))
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -401,7 +401,7 @@ func TestSessionNewSenderDuplicateLinks(t *testing.T) {
 func TestSessionNewSenderMaxHandles(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(SenderSettleModeUnsettled))
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -432,7 +432,7 @@ func TestSessionNewSenderMaxHandles(t *testing.T) {
 func TestSessionUnexpectedFrame(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(SenderSettleModeUnsettled))
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -455,7 +455,7 @@ func TestSessionUnexpectedFrame(t *testing.T) {
 func TestSessionInvalidFlowFrame(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(SenderSettleModeUnsettled))
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -513,7 +513,7 @@ func TestSessionFlowFrameWithEcho(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -558,7 +558,7 @@ func TestSessionInvalidAttachDeadlock(t *testing.T) {
 		}
 	}
 	netConn := mocks.NewNetConn(responder)
-	client, err := New(netConn, nil)
+	client, err := NewConn(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
